requirements.txt
streamlit==1.55.0
google-cloud-firestore>=2.17.0
google-api-python-client>=2.151.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
googlemaps>=4.10.0
pandas>=2.2.0
plotly>=5.22.0
jpholiday>=1.0.0

